fix: refactoring and fix for test
