+++ packages/pn-personafisica-webapp/src/component/notification/__test__/MobileNotifications.test.tsx
@@ -1,4 +1,4 @@
-import { render }  from "@testing-library/react";
+import { render } from "../../../__test__/test-utils";
 import MobileNotifications from "../MobileNotifications";
 
 jest.mock('@pagopa-pn/pn-commons', () => {
